rafactor: reader
